Split globus_app into multiple modules
