+++ src/globus_sdk/experimental/globus_app/globus_app.py
@@ -1,143 +1,21 @@
 from __future__ import annotations
 
 import abc
-import dataclasses
-import sys
-import typing as t
-from dataclasses import dataclass
 
-from globus_sdk import (
-    AuthClient,
-    AuthLoginClient,
-    ConfidentialAppAuthClient,
-    NativeAppAuthClient,
-    Scope,
-)
-from globus_sdk import config as sdk_config
+from globus_sdk import AuthClient, AuthLoginClient, GlobusSDKUsageError, Scope
 from globus_sdk._types import ScopeCollectionType, UUIDLike
 from globus_sdk.authorizers import GlobusAuthorizer
-from globus_sdk.exc import GlobusSDKUsageError
 from globus_sdk.experimental.auth_requirements_error import (
     GlobusAuthorizationParameters,
 )
-from globus_sdk.experimental.login_flow_manager import (
-    CommandLineLoginFlowManager,
-    LocalServerLoginFlowManager,
-    LoginFlowManager,
-)
-from globus_sdk.experimental.tokenstorage import (
-    JSONTokenStorage,
-    MemoryTokenStorage,
-    SQLiteTokenStorage,
-    TokenStorage,
-)
+from globus_sdk.experimental.tokenstorage import TokenStorage
 from globus_sdk.scopes import AuthScopes, scopes_to_scope_list
 
+from ._types import TokenStorageProvider
 from ._validating_token_storage import ValidatingTokenStorage
-from .authorizer_factory import (
-    AccessTokenAuthorizerFactory,
-    AuthorizerFactory,
-    ClientCredentialsAuthorizerFactory,
-    RefreshTokenAuthorizerFactory,
-)
-from .errors import IdentityMismatchError, TokenValidationError
-
-if sys.version_info < (3, 8):
-    from typing_extensions import Protocol, runtime_checkable
-else:
-    from typing import Protocol, runtime_checkable
-
-
-@runtime_checkable
-class TokenStorageProvider(Protocol):
-    @classmethod
-    def for_globus_app(
-        cls, client_id: UUIDLike, app_name: str, config: GlobusAppConfig, namespace: str
-    ) -> TokenStorage: ...
-
-
-@runtime_checkable
-class LoginFlowManagerProvider(Protocol):
-    @classmethod
-    def for_globus_app(
-        cls, app_name: str, login_client: AuthLoginClient, config: GlobusAppConfig
-    ) -> LoginFlowManager: ...
-
-
-class TokenValidationErrorHandler(Protocol):
-    def __call__(self, app: GlobusApp, error: TokenValidationError) -> None: ...
-
-
-def resolve_by_login_flow(app: GlobusApp, error: TokenValidationError) -> None:
-    """
-    An error handler for GlobusApp token access errors that will retry the
-    login flow if the error is a TokenValidationError.
-
-    :param app: The GlobusApp instance which encountered an error.
-    :param error: The encountered token validation error.
-    """
-    if isinstance(error, IdentityMismatchError):
-        # An identity mismatch error indicates incorrect use of the app. Not something
-        #   that can be resolved by running a login flow.
-        raise error
-
-    app.run_login_flow()
-
-
-KnownLoginFlowManager = t.Literal["command-line", "local-server"]
-KNOWN_LOGIN_FLOW_MANAGERS: dict[KnownLoginFlowManager, LoginFlowManagerProvider] = {
-    "command-line": CommandLineLoginFlowManager,
-    "local-server": LocalServerLoginFlowManager,
-}
-
-KnownTokenStorage = t.Literal["json", "sqlite", "memory"]
-KNOWN_TOKEN_STORAGES: dict[KnownTokenStorage, t.Type[TokenStorageProvider]] = {
-    "json": JSONTokenStorage,
-    "sqlite": SQLiteTokenStorage,
-    "memory": MemoryTokenStorage,
-}
-
-
-@dataclass(frozen=True)
-class GlobusAppConfig:
-    """
-    Various configuration options for controlling the behavior of a ``GlobusApp``.
-
-    :param login_flow_manager: An optional ``LoginFlowManager`` instance, provider,
-        or identifier ("command-line" or "local-server").
-        For a ``UserApp``, defaults to "command-line".
-        For a ``ClientApp``, this value is not supported.
-    :param login_redirect_uri: The redirect URI to use for login flows.
-        For a "local-server" login flow manager, this value is not supported.
-        For a native client, this value defaults to a globus-hosted helper page.
-        For a confidential client, this value is required.
-    :param request_refresh_tokens: If True, the ``GlobusApp`` will request refresh
-        tokens for long-lived access.
-    :param token_storage: A ``TokenStorage`` instance, provider, or identifier
-        ("json", "sqlite", or "memory").
-        Default: "json"
-    :param token_validation_error_handler: A callable that will be called when a
-        token validation error is encountered. The default behavior is to retry the
-        login flow automatically.
-    :param environment: The Globus environment being targeted by this app. This is
-        predominately for internal use and can be ignored in most cases.
-    """
-
-    login_flow_manager: (
-        KnownLoginFlowManager | LoginFlowManagerProvider | LoginFlowManager | None
-    ) = None
-    login_redirect_uri: str | None = None
-    token_storage: KnownTokenStorage | TokenStorageProvider | TokenStorage = "json"
-    request_refresh_tokens: bool = False
-    token_validation_error_handler: TokenValidationErrorHandler | None = (
-        resolve_by_login_flow
-    )
-    environment: str = dataclasses.field(
-        default_factory=sdk_config.get_environment_name
-    )
-
-
-_DEFAULT_CONFIG = GlobusAppConfig()
+from .authorizer_factory import AuthorizerFactory
+from .config import DEFAULT_CONFIG, KNOWN_TOKEN_STORAGES, GlobusAppConfig
+from .errors import TokenValidationError
 
 
 class GlobusApp(metaclass=abc.ABCMeta):
@@ -166,7 +44,7 @@ class GlobusApp(metaclass=abc.ABCMeta):
         client_id: UUIDLike | None = None,
         client_secret: str | None = None,
         scope_requirements: dict[str, ScopeCollectionType] | None = None,
-        config: GlobusAppConfig = _DEFAULT_CONFIG,
+        config: GlobusAppConfig = DEFAULT_CONFIG,
     ):
         """
         :param app_name: A string to identify this app. Used for default tokenstorage
@@ -413,232 +291,3 @@ class GlobusApp(metaclass=abc.ABCMeta):
         :param resource_server: the resource server to get scope requirements for
         """
         return tuple(self._scope_requirements.get(resource_server, []))
-
-
-class UserApp(GlobusApp):
-    """
-    A ``GlobusApp`` for login methods that require an interactive flow with a user.
-
-    ``UserApp``\\s are most commonly used with native application clients by passing a
-    ``NativeAppAuthClient`` as ``login_client`` or the native application's
-    ``client_id``.
-
-    If using a templated client, either pass a ``ConfidentialAppAuthClient``
-    as `login_client`` or the templated client's ``client_id`` and ``client_secret``.
-    This will not work for standard confidential clients.
-
-    By default a ``UserApp`` will create a ``CommandLineLoginFlowManager`` for
-    running login flows, which can be overridden through ``config``.
-
-    .. tab-set::
-
-        .. tab-item:: Example Usage
-
-            .. code-block:: python
-
-                app = UserApp("myapp", client_id=NATIVE_APP_CLIENT_ID)
-                client = TransferClient(app=app)
-                res = client.endpoint_search("Tutorial Collection")
-
-    """
-
-    _login_client: NativeAppAuthClient | ConfidentialAppAuthClient
-    _authorizer_factory: (  # type:ignore
-        AccessTokenAuthorizerFactory | RefreshTokenAuthorizerFactory
-    )
-
-    def __init__(
-        self,
-        app_name: str = "Unnamed Globus App",
-        *,
-        login_client: AuthLoginClient | None = None,
-        client_id: UUIDLike | None = None,
-        client_secret: str | None = None,
-        scope_requirements: dict[str, ScopeCollectionType] | None = None,
-        config: GlobusAppConfig = _DEFAULT_CONFIG,
-    ):
-        super().__init__(
-            app_name,
-            login_client=login_client,
-            client_id=client_id,
-            client_secret=client_secret,
-            scope_requirements=scope_requirements,
-            config=config,
-        )
-
-        self._login_flow_manager = self._resolve_login_flow_manager(
-            app_name=self.app_name,
-            login_client=self._login_client,
-            config=config,
-        )
-
-    def _resolve_login_flow_manager(
-        self, app_name: str, login_client: AuthLoginClient, config: GlobusAppConfig
-    ) -> LoginFlowManager:
-        login_flow_manager = config.login_flow_manager
-        if isinstance(login_flow_manager, LoginFlowManager):
-            return login_flow_manager
-
-        elif isinstance(login_flow_manager, LoginFlowManagerProvider):
-            provider = login_flow_manager
-        elif login_flow_manager is None:
-            provider = CommandLineLoginFlowManager
-        elif login_flow_manager in KNOWN_LOGIN_FLOW_MANAGERS:
-            provider = KNOWN_LOGIN_FLOW_MANAGERS[login_flow_manager]
-        else:
-            allowed_keys = ", ".join(repr(k) for k in KNOWN_LOGIN_FLOW_MANAGERS.keys())
-            raise GlobusSDKUsageError(
-                f"Unsupported login_flow_manager value: {login_flow_manager!r}. "
-                f"Expected {allowed_keys}, a <LoginFlowManagerProvider>, or a "
-                f"<LoginFlowManager>."
-            )
-
-        return provider.for_globus_app(app_name, login_client, config)
-
-    def _initialize_login_client(
-        self,
-        app_name: str,
-        config: GlobusAppConfig,
-        client_id: UUIDLike,
-        client_secret: str | None,
-    ) -> AuthLoginClient:
-        if client_secret:
-            return ConfidentialAppAuthClient(
-                app_name=app_name,
-                client_id=client_id,
-                client_secret=client_secret,
-                environment=config.environment,
-            )
-        else:
-            return NativeAppAuthClient(
-                app_name=app_name,
-                client_id=client_id,
-                environment=config.environment,
-            )
-
-    def _initialize_authorizer_factory(self) -> None:
-        if self.config.request_refresh_tokens:
-            self._authorizer_factory = RefreshTokenAuthorizerFactory(
-                token_storage=self._validating_token_storage,
-                auth_login_client=self._login_client,
-            )
-        else:
-            self._authorizer_factory = AccessTokenAuthorizerFactory(
-                token_storage=self._validating_token_storage,
-            )
-
-    def run_login_flow(
-        self, auth_params: GlobusAuthorizationParameters | None = None
-    ) -> None:
-        """
-        Run an authorization flow to get new tokens which are stored and available
-        for the next authorizer gotten by get_authorizer.
-
-        As a UserApp this always involves an interactive login flow with the user
-        driven by the app's LoginFlowManager.
-
-        :param auth_params: A GlobusAuthorizationParameters to control how the user
-            will authenticate. If not passed
-        """
-        auth_params = self._auth_params_with_required_scopes(auth_params)
-        token_response = self._login_flow_manager.run_login_flow(auth_params)
-        self._authorizer_factory.store_token_response_and_clear_cache(token_response)
-
-
-class ClientApp(GlobusApp):
-    """
-    A ``GlobusApp`` using client credentials - useful for service accounts and
-    automation use cases.
-
-    ``ClientApp``\\s are always used with confidential clients either by passing
-    a ``ConfidentialAppAuthClient`` as ``login_client`` or providing the client's
-    ``client_id`` and ``client_secret`` pair.
-
-    ``ClientApp``\\s do not use a ``LoginFlowManager`` and will raise an error
-    if given one through ``config``.
-
-    .. tab-set::
-
-        .. tab-item:: Example Usage
-
-            .. code-block:: python
-
-                app = ClientApp("myapp", CLIENT_ID, CLIENT_SECRET)
-                client = TransferClient(app=app)
-                res = client.endpoint_search("Tutorial Collection")
-
-    """
-
-    _login_client: ConfidentialAppAuthClient
-    _authorizer_factory: ClientCredentialsAuthorizerFactory  # type:ignore
-
-    def __init__(
-        self,
-        app_name: str = "Unnamed Globus App",
-        *,
-        login_client: ConfidentialAppAuthClient | None = None,
-        client_id: UUIDLike | None = None,
-        client_secret: str | None = None,
-        scope_requirements: dict[str, ScopeCollectionType] | None = None,
-        config: GlobusAppConfig = _DEFAULT_CONFIG,
-    ):
-        if config.login_flow_manager is not None:
-            raise GlobusSDKUsageError("A ClientApp cannot use a login_flow_manager")
-
-        if login_client and not isinstance(login_client, ConfidentialAppAuthClient):
-            raise GlobusSDKUsageError(
-                "A ClientApp must use a ConfidentialAppAuthClient for its login_client"
-            )
-
-        super().__init__(
-            app_name,
-            login_client=login_client,
-            client_id=client_id,
-            client_secret=client_secret,
-            scope_requirements=scope_requirements,
-            config=config,
-        )
-
-    def _initialize_login_client(
-        self,
-        app_name: str,
-        config: GlobusAppConfig,
-        client_id: UUIDLike,
-        client_secret: str | None,
-    ) -> AuthLoginClient:
-        if not client_secret:
-            raise GlobusSDKUsageError(
-                "A ClientApp requires a client_secret to initialize its own login "
-                "client."
-            )
-
-        return ConfidentialAppAuthClient(
-            client_id=client_id,
-            client_secret=client_secret,
-            app_name=app_name,
-            environment=config.environment,
-        )
-
-    def _initialize_authorizer_factory(self) -> None:
-        self._authorizer_factory = ClientCredentialsAuthorizerFactory(
-            token_storage=self._validating_token_storage,
-            confidential_client=self._login_client,
-        )
-
-    def run_login_flow(
-        self, auth_params: GlobusAuthorizationParameters | None = None
-    ) -> None:
-        """
-        Run an authorization flow to get new tokens which are stored and available
-        for the next authorizer gotten by get_authorizer.
-
-        As a ClientApp this is just a client credentials call.
-
-        :param auth_params: A GlobusAuthorizationParameters to control authentication
-            only the required_scopes parameter is used.
-        """
-        auth_params = self._auth_params_with_required_scopes(auth_params)
-        token_response = self._login_client.oauth2_client_credentials_tokens(
-            requested_scopes=auth_params.required_scopes
-        )
-        self._authorizer_factory.store_token_response_and_clear_cache(token_response)
